wifi: brcmfmac: Add BCM43342/1 support
This chip is found in Rinato, and some chinese set-top boxes as well.
Most of the patch is taken from the Armbian forums.

Scanning for Wi-Fi networks works, but connecting doesn't.
This may be caused by missing a firmware file for these chips
(the so-called clm_blob, containing regulatory information).